Created basic elevator subsystem
I copied the shooter subsystem and updated it so that it will work for the elevator system instead. Still need the elevator motor can ID and to set the min and max for the elevator.  In general 1 file created and set up as well as putting notes in constants for clearing up why certain error are there.